+++ directio-project/asakusa-directio-dmdl/src/main/resources/com/asakusafw/dmdl/directio/text/messages.properties
@@ -12,6 +12,7 @@ AttributeAnalyzer.diagnosticNotDecimalFormat={0} must be a valid pattern string 
 AttributeAnalyzer.diagnosticNotEnumConstant={0} must be one of '{'{1}'}'
 AttributeAnalyzer.diagnosticNotMap={0} must be a map
 AttributeAnalyzer.diagnosticNotString={0} must be a string
+AttributeAnalyzer.diagnosticNotTimeZoneId={0} must be a valid time zone ID
 EscapeSettings.diagnosticConflictLineSeparator={0} must not be line separator characters
 EscapeSettings.diagnosticEscapeSequenceDuplicateKey={0} has duplicated key: {1}
 EscapeSettings.diagnosticEscapeSequenceDuplicateValue={0} has duplicated value: {1}
